bug fixes in register route

+++ src/utills/cloudinary.js
@@ -7,12 +7,14 @@ cloudinary.config({
 })
 const uploadOnCloudinary=async (localFilePath)=>{
     try {
+        if(!localFilePath) return null
         const uploadResult= await cloudinary.uploader.upload(localFilePath,{
-            resource_type:auto
+            resource_type: "auto"
         })
         fs.unlinkSync(localFilePath)
         return uploadResult
     } catch (error) {
+        console.log(error)
         fs.unlinkSync(localFilePath)
         return null
     }
